Fix fatal error filterColorRed not found and filterBrightness not found

+++ Plugins/Renderer/gl_shader.cpp
@@ -369,7 +369,7 @@ void R_SaveProgramStatesCaches(const char *filename, const std::vector<program_s
 		}
 		else
 		{
-			for (int i = 0; i < mapping_size; ++i)
+			for (size_t i = 0; i < mapping_size; ++i)
 			{
 				if (p & mapping[i].state)
 				{
